requirements.txt
# Core Flask Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3

# Database & ORM
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
psycopg2-binary==2.9.9

# Authentication & Authorization
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0

# Email
Flask-Mail==0.9.1

# Environment & Configuration
python-dotenv==1.0.0

# Web Server (Production)
gunicorn==21.2.0
python-socketio==5.10.0

# Caching & Task Queue
Flask-Caching==2.1.0
celery==5.3.4
redis==5.0.1

# API & Utilities
requests==2.31.0
python-dateutil==2.8.2

# Data Validation & Serialization
marshmallow==3.20.1
python-dateutil==2.8.2

# File Uploads
python-magic==0.4.27

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Code Quality
flake8==6.1.0
black==23.12.0
isort==5.13.2
bandit==1.7.5
safety==2.3.5

# Deployment
vercel==1.0.0

# Security
cryptography==41.0.7
bleach==6.1.0
